The CALQ decoder now accepts input files with any file extension; finished run_calq.sh

+++ evaluation/scripts/run_codecs/run_calq.sh
@@ -26,41 +26,53 @@ printf "OK\n"
 
 time="/usr/bin/time"
 python="/usr/bin/python"
-calq="/project/dna/install/calq-d3965e2/calq"
+calq="/home/voges/git/calq/build/calq"
 calq_string="calq-d3965e2"
+grep="/usr/bin/grep"
+ps="/usr/bin/ps"
+awk="/usr/bin/awk"
 replace_qual_sam_py="/home/voges/git/ngstools/replace_qual_sam.py"
+ps_mem_py="/home/voges/bin/ps_mem.py"
 
 printf "Checking executables ... "
 if [ ! -x $time ]; then printf "did not find $time\n"; exit -1; fi
 if [ ! -x $python ]; then printf "did not find $python\n"; exit -1; fi
 if [ ! -x $calq ]; then printf "did not find $calq\n"; exit -1; fi
 if [ ! -e $replace_qual_sam_py ]; then printf "did not find $replace_qual_sam_py\n"; exit -1; fi
+if [ ! -e $ps_mem_py ]; then printf "did not find $ps_mem_py\n"; exit -1; fi
 printf "OK\n"
 
 ###############################################################################
 #                          Compress and decompress                            #
 ###############################################################################
 
-printf "[1/3] Running CALQ encoder ... "
-$time -v -o $input_sam.$calq_string.enc.time $calq -f -q Illumina-1.8+ -p $polyploidy -b $block_size $input_sam -o $input_sam.cq &> $input_sam.$calq_string.enc.log
-printf "OK\n"
-
-printf "[2/3] Running CALQ decoder ... "
-$time -v -o $input_sam.$calq_string.dec.time $calq -f -d -s $input_sam $input_sam.cq &> $input_sam.$calq_string.dec.log
-printf "OK\n"
-
-printf "[3/3] Constructing SAM file with reconstruced quality values ... "
+printf "Running CALQ encoder ... "
+cmd="$calq -f -q Illumina-1.8+ -p $polyploidy -b $block_size $input_sam -o $input_sam.cq"
+$time -v -o $input_sam.$calq_string.enc.time $cmd &> $input_sam.$calq_string.enc.log &
+pid=$($ps aux | $grep "$cmd" | $grep -v grep | $grep -v "$time" | $awk '{print $2}')
+printf "Command being traced: \"$cmd\"\n" > $input_sam.$calq_string.enc.mem
+$python $ps_mem_py -w 1 --swap -p $pid >> $input_sam.$calq_string.enc.mem
 mv $input_sam.cq $input_sam.$calq_string
-mv $input_sam.cq.qual $input_sam.$calq_string.qual
-$python $replace_qual_sam_py $input_sam $input_sam.$calq_string.qual &> $input_sam.run_calq.log
-mv $input_sam.new_qual.sam $input_sam.$calq_string.sam
 printf "OK\n"
+
+printf "Running CALQ decoder ... "
+cmd="$calq -f -d -s $input_sam $input_sam.$calq_string -o $input_sam.$calq_string.qual"
+$time -v -o $input_sam.$calq_string.dec.time $cmd &> $input_sam.$calq_string.dec.log &
+pid=$($ps aux | $grep "$cmd" | $grep -v grep | $grep -v "$time" | $awk '{print $2}')
+printf "Command being traced: \"$cmd\"\n" > $input_sam.$calq_string.dec.mem
+$python $ps_mem_py -w 1 --swap -p $pid >> $input_sam.$calq_string.dec.mem
+printf "OK\n"
+
+#printf "Constructing SAM file with reconstruced quality values ... "
+#$python $replace_qual_sam_py $input_sam $input_sam.$calq_string.qual &> $input_sam.run_calq.log
+#mv $input_sam.new_qual.sam $input_sam.$calq_string.sam
+#printf "OK\n"
 
 ###############################################################################
 #                                   Cleanup                                   #
 ###############################################################################
 
 printf "Cleanup ... "
-#
+rm $input_sam.$calq_string.qual
 printf "OK\n";
